libbpf-tools/filelife: Fix error tracking caused by failed unlink
When we try to delete a file with a normal user after creating a file as
root (sudo), the deletion of the file fails, but the kprobe:vfs_unlink is
triggered, which leads to incorrect file lifecycle tracking.

This commit makes the same changes [1] as tools/filelife.py, or you can
see bcc [2] commit 076ccf0 ("tools/filelife: Fix error tracking caused
by failed unlink").

[1] #4791
[2] 076ccf0

Signed-off-by: Rong Tao <rongtao@cestc.cn>
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I13e424b5025ac3d564bf650c20dcc08276dca92f
Reviewed-on: http://gerrit.kylin.com/c/ketones/+/82652