libbpf-tools: Fix misaligned pointer accesses in opensnoop
The perf buffer in opensnoop doesn't maintain 8 byte alignment
for ts and callers in struct event. When building with "
-fsanitize=alignment -fsanitize-trap=undefined" failures happen
in handle_event. Fix these by copying the event from the perf
buffer before accessing.

This is similar to a fix in exitsnoop where different ways to handle
misaligned pointers were discussed:

Signed-off-by: Ian Rogers <irogers@google.com>
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I330dd6e624edcc54b8ee738802d876a623aba30d
Reviewed-on: http://gerrit.kylin.com/c/ketones/+/94821