Fix bashreadline
This Pull Request fixes the issue where the bashreadline tool would
lose the first bash command upon startup.

At runtime, bash blocks at the readline function in bash::readline.c
Upon receiving the Enter key signal, it calls the readline_internal_teardown
function to retrieve the user's input. Therefore, changing the hook
function from readline to readline_internal_teardown resolves the issue
of losing the first bash command after startup of bashreadline.

The specific approach is to open the symbol table of bash and check
if there is a symbol for readline_internal_teardown function. If not,
continue to use the readline function as the hook position.

Author: @sancppp
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I88e26ffd69be7daae72aaad2266770ffffa332bc
Reviewed-on: http://gerrit.kylin.com/c/ketones/+/99304