libbpf-tools/bitesize.bpf.c: Fix potential out-of-bounds access in comm_allowed function
fixes a potential out-of-bounds access in the comm_allowed function. Previously, the condition
in the for loop checked the value of targ_comm[i] before ensuring that i is less than TASK_COMM_LEN.
This could lead to out-of-bounds access if i equals TASK_COMM_LEN.

The condition in the for loop has been updated to check that i is less than TASK_COMM_LEN
before accessing targ_comm[i]. This ensures that targ_comm is not accessed out-of-bounds.

@chudihuang
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I61bc9adbb9539215284abfc61a65bc2388357952
Reviewed-on: http://gerrit.kylin.com/c/ketones/+/134631